Remove ssh tcp check.
Kitchen docker uses the "wait_for_sshd" method in Test Kitchen to
determine when the container is available to ssh into. Since the Docker
server builds containers with SSH access on random ports, this presents
a problem when on networks that deny TCP checks to these ports.

As a workaround, I implemented an alternative that checks the state of
the Docker container to determine if it is up.