Override 'wait_for_sshd' method
Kitchen's `ssh_base.wait_for_sshd` methods does a check whether the port
is listening and ready. But docker forwards the connection to a
container's internal port. `ssh_base.wait_for_sshd` returns that is
always ready even if inside the container the ssh service is not.

We need to perform an actual connection using the ssh client to make
sure the service is ready and hence, this patch overrides the
`wait_for_sshd` method.