support passing in '-e' flags - fixes systemd init
This commit allows for passing in '-e' flags in the driver config. Passing in an '-e' flag for container=docker allows systems such as CentOS7, which use systemd, to run systemd within the container. This allows for testing when cookbooks install and start services on servers which use systemd as their init system. This can be verified using image: centos:7 , and the following driver_config:

    driver_config:
      privileged: true
      run_command: /usr/sbin/init
      volume:
        - "/sys/fs/cgroup:/sys/fs/cgroup:ro"
      provision_command: systemctl enable sshd.service
      eflags: container=docker