env container=docker for debians, ext #192
When running /sbin/init from systemd in debian:stretch (recent, unable
to determine since when), it is necessary to include the enviroment
variable container=docker in the Dockerfiles. This is essentially the
same as for recent CentOSes as previously fixed in #192.

Tests with ubuntu-14.04 have shown no hurts and I strongly suspect the
same for other debian-based images.

	modified:   lib/kitchen/driver/docker.rb