added ssh_host option
Support for playing nicely with SSH transport's `ssh_gateway` option. If
set, kitchen ssh will attempt to connect to the SSH daemon running on the
specified host. This allows for running tests against remote docker engines
without the need to open additional ports on the remote host. Usually you will
want to set this to `localhost` when used in conjunction with SSH
transport's `ssh_gateway` option.

Support for playing nicely with SSH transport