Add ability to use ~ in public and private key path
Having the ability to use the `~` in `public_key` and `private_key`
which gets evaluated to the current users home directory can be useful
in preventing several different images from being created. Each time
kitchen-docker runs it looks for existing SSH keys and will use them.
When using Docker cache this means no new images is created. If no keys
are found they are randomly generated and as a result, a new image is
created.